refactor: rename Email component
