+++ components/Email/index.ts
@@ -1,0 +1,1 @@
+export { default as Email } from "./Email";
